Add simple UI test using mock server
This uses the full application code, but mocks out the base URL for requests so that they'll go to the mock server instead of a real API. This helps make the tests more predictable/less flaky/more hermetic.

There were a few quirks though, especially around dependency version conflicts as well as adding a thing to the manifest to allow clear text traffic for the mock server to work. Maybe there's a good way to have this only available in the application for tests though (or maybe at least just debug builds)? It seems like a potential security risk somehow (although I don't really understand the specifics well enough to know)