Remove unnecessary runBlocking call
The ViewModel didn't really need to (and probably shouldn't have) used runBlocking to trigger the network request. If I understand correctly, this was actually blocking the main thread while the response was waiting to come back.

A consequence of this change though is that the tests for this no longer passed. This is because the assertions on the view were now happening before the UI updated (unlike before where the main thread was blocked and therefore the Espresso assertions waited). I as able to fix this by using a call in the test that waits until the request has been finished before continuing, so that I can be sure that things are done before making view assertions.

Another change that I made here was switching from using postValue to using setValue (well, more specifically, the property accessor `value`). I'm able to do this because the coroutine scope is started on the main dispatcher, and even though the API call happens on a background dispatcher (because it was implemented that way in Retrofit), when it finishes, the call to update the LiveData will happen on the main thread still (because this is main-safe - see this article for more details: https://medium.com/androiddevelopers/coroutines-on-android-part-i-getting-the-background-3e0e54d20bb ). Becuase this is already happening on the main thread, I can use setValue instead of posting.