Add dynamic values for buttons on results screen
This uses Firebase Remote config to determine the values.

There are default values as well built-in to the app just in case the remote values aren't available.

I also tried to encapsulate the logic of interacting with Firebase so that dependencies on it are as minimal as possible. I could probably take this even further in the Application class by extracting the firebase stuff into a separate class, but it's probably overkill for the moment.