Ignore failures from detekt
I was attempting to use the `ignoreFailures` flag, but as the new comment states, this is not available in a public repository yet (and I don't want to build it myself). You can see the PR where this was added (9 days ago as of this commit) here: https://github.com/arturbosch/detekt/pull/1600. Hopefull ythis il be in the next version though.

In the meantime, I've gone about ignoring errors by adding a config file and setting the maxIssues value to -1 (as suggested here: https://github.com/arturbosch/detekt/issues/1435). This effectively ignores all issues for build purposes. I want this for now so that at least my CI builds can pass. I want this stuff mainly for informational purposes, not for enforcement.