Use new OkHttp rule for MainActivity tests
This handles idling, so I can remove the old wait method.

This also caused a slight change in one of the tests. In the "go to questions screen then press back" test, I wasn't actually using a response (which was a bit hacky, perhaps) and just pressing back immediately. However, with the new idling resource, the test idles waiting for a response that is never coming, so it fails. I ended up updating things to just always use a response before starting the question screen to fix this.