Use correct version of JaCoCo for Android tests
I was running into a weird issue where Kotlin generated code was not being properly ignored, which it should have been in version 0.8.2+. I later noticed that my JaCoCo report files were showing an older version of JaCoCo than what I had specified. it turns out that you have to specify the version separately for the Android block (which is the opposite of what one of the articles had me believe). Carmen A. helped me out with this in ASG, so props!