Attempt to ignore more files
I don't really want my DI setup checked for code coverage, nor do I want generated navigation directions counted either.

I say "attempt to" because I tried this on my local machine, but the excluded stuff still showed up. But I have strong reason to believe it is because of a caching issue that I can't quite figure out. So, committing to get CI to build it to see if it makes a difference.