Use injected preferences in Application class
Since this is injectable now, it makes sense to just inject it in the application where it was being used.

This did mean that I needed to move Koin startup to before I set the theme though. This may ultimately be undesirable in the case that it cases theme setup to be too slow, but not really an issue for my purposes for the most part.